Update：更新CF，添加网络消息处理工具；

+++ GameServer/Framework/CosmosFramework/Core/RunTime/Network/INetworkMessage.cs
@@ -7,7 +7,7 @@ namespace Cosmos
     /// </summary>
     public interface INetworkMessage
     {
-        uint Conv { get; }
+        long Conv { get; }
         byte[] EncodeMessage();
         bool DecodeMessage(byte[] buffer);
         byte[] GetBuffer();
